requirements.txt
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
datawrapper==0.4.6
Deprecated==1.2.13
google==3.0.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gspread==5.7.2
httplib2==0.20.4
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==4.0.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
